Fixed unit test
